Refactor some logic about merge ops out of UsfmEditor
